Move DashboardContentShelf component to its own file
